# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  models.AdjustStockRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      reason:
        type: string
      reference:
        type: string
      type:
        enum:
        - in
        - out
        - adjustment
        type: string
    required:
    - product_id
    - quantity
    - reason
    - type
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateInventoryRequest:
    properties:
      location:
        type: string
      product_id:
        type: string
      quantity:
        minimum: 0
        type: integer
    required:
    - location
    - product_id
    - quantity
    type: object
...
  models.Inventory:
    properties:
      created_at:
        type: string
      id:
        type: string
      location:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  models.InventoryTransaction:
    properties:
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        description: changed from uuid.UUID to string
        type: string
      quantity:
        type: integer
      reason:
        type: string
      reference:
        type: string
      type:
        description: '"in", "out", "adjustment"'
        type: string
    type: object
...
  models.Product:
    properties:
      barcode_number:
        type: string
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      updated_at:
        type: string
    type: object
...
  models.UpdateInventoryRequest:
    properties:
      location:
        type: string
      quantity:
        minimum: 0
        type: integer
    required:
    - location
    - quantity
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@jatistore.local
    name: API Support
  description: RESTful API for Point of Sales (POS) system using Go, Fiber, and PostgreSQL.
    Includes inventory management, customer management, order processing, payments,
    and receipts.
  title: JatiStore POS API
  version: "2.0"
paths:
...
  /inventory:
    get:
      consumes:
      - application/json
      description: Get a list of all inventory records
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Inventory'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all inventory records
      tags:
      - Inventory
    post:
      consumes:
      - application/json
      description: Create a new inventory record with the provided data
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Inventory data
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/models.CreateInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Inventory'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a new inventory record
      tags:
      - Inventory
  /inventory/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an inventory record by its ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete an inventory record
      tags:
      - Inventory
    get:
      consumes:
      - application/json
      description: Get an inventory record by its unique identifier
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Inventory'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get inventory by ID
      tags:
      - Inventory
    put:
      consumes:
      - application/json
      description: Update an inventory record with the provided data
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Inventory ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated inventory data
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/models.UpdateInventoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Inventory'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Update an inventory record
      tags:
      - Inventory
  /inventory/adjust:
    post:
      consumes:
      - application/json
      description: Adjust inventory stock levels (in/out/adjustment)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Stock adjustment data
        in: body
        name: adjustment
        required: true
        schema:
          $ref: '#/definitions/models.AdjustStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.InventoryTransaction'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Adjust inventory stock
      tags:
      - Inventory
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"